refactor: `DTO` 위치 이동
